requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
async-timeout==5.0.1
attrs==25.3.0
banal==1.0.6
biopython==1.85
causal-conv1d==1.5.0.post8
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==3.5.0
dill==0.3.8
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.9
mamba-ssm==2.2.4
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.4
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
propcache==0.3.1
psutil==7.0.0
pyarrow==19.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
SQLAlchemy==1.4.54
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
triton==3.2.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
